src/backend/requirements.txt
python-dotenv==1.2.2
sqlalchemy==2.0.49
fastapi==0.135.3
pydantic==2.12.5

# other
starlette==1.0.0 # needed for middleware cros support in fastapi    